convert to package structure
